# backend/src/main/resources/openapi/sims-openapi.yaml
openapi: 3.0.3
info:
  title: SIMS API — Exam, Fee and Communication modules
  description: |
    School Information Management System — the exam, fee and communication modules
    after the business-workflow redesign.

    ## Conventions

    **Nested business objects.** Responses embed the data a screen needs rather than
    bare foreign keys: `subject`, `student`, `schoolClass`, `createdBy` and
    `updatedBy` are objects carrying ids *and* names. The old flat `*Id` fields are
    still returned but are deprecated and will be removed once the React app has
    migrated off them.

    **Pagination.** List endpoints return a Spring `Page`:
    `?page=0&size=20&sort=createdAt,desc`.

    **Actor fields.** Requests carry `actorId` (admin staff) or `enteredById` /
    `updatedById` (teacher) because authentication is still being built by another
    developer. These fields disappear once the JWT principal is available.

    **Marks and fee "sheets".** Marks and fee statuses are worked on a section at a
    time, so each has a `GET /sheet` that returns the whole roster pre-filled and a
    `PUT /sheet` that saves it in one transaction.
  version: 2.0.0
  contact:
    name: SIMS Team
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/exams:
    get:
      tags: [Exams]
      summary: List exams
      description: Paginated, optionally filtered by academic year and/or status.
      operationId: getAllExams
      parameters:
        - $ref: '#/components/parameters/AcademicYearIdFilter'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ExamStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of exams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageExamResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Exams]
      summary: Create an exam
      description: Status defaults to `DRAFT`; publishing is done via the publish endpoint.
      operationId: createExam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamRequest'
      responses:
        '200':
          description: Exam created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: An exam with this name already exists for the academic year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exams/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exams]
      summary: Get exam by id
      operationId: getExamById
      responses:
        '200':
          description: Exam
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Exams]
      summary: Update an exam
      description: |
        Rejected once the exam is `PUBLISHED`. The status field is ignored here —
        status only changes through the publish endpoint.
      operationId: updateExam
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamRequest' }
      responses:
        '200':
          description: Exam updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Exam is published, or the name duplicates another exam in the year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Exams]
      summary: Delete an exam
      description: |
        Rejected when the exam is published, when marks have been recorded against
        it, or while subjects are still mapped to it.
      operationId: deleteExam
      responses:
        '204': { description: Exam deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Published, has marks, or still has subjects mapped
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exams/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/PathId'
    put:
      tags: [Exams]
      summary: Publish an exam
      description: |
        Rejected unless every precondition holds: at least one subject is mapped and
        every mapped subject has a timetable slot. Call `publish-preview` first to
        show the user what is missing.
      operationId: publishExam
      responses:
        '200':
          description: Exam published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already published, or preconditions not met
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                timestamp: '2026-08-02T10:30:00Z'
                status: 409
                error: Conflict
                message: 'Exam cannot be published: 2 subject(s) have not been scheduled in the timetable.'
                path: /api/v1/exams/9f1c.../publish
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exams/{id}/overview:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exams]
      summary: Exam dashboard
      description: |
        Exam header plus every mapped subject with its schedule and marks-entry
        progress — the whole dashboard in one call. Marks counts come from a single
        grouped query, not one query per subject.
      operationId: getExamOverview
      responses:
        '200':
          description: Exam overview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamOverviewResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exams/{id}/publish-preview:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exams]
      summary: Publish readiness checklist
      description: |
        Reports what blocks publishing without attempting it, so the UI can render a
        checklist instead of guessing from an error message.
      operationId: getExamPublishPreview
      responses:
        '200':
          description: Readiness report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishPreviewResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Exam Subjects
  # ==================================================================
...
  /api/v1/exam-subjects/exam/{examId}:
    parameters:
      - name: examId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Exam Subjects]
      summary: Subjects mapped to an exam
      description: Bounded by the exam, so returned unpaginated.
      operationId: getExamSubjectsByExamId
      responses:
        '200':
          description: Exam subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExamSubjectResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Exam Timetable
  # ==================================================================
...
  /api/v1/exam-timetables/exam/{examId}:
    parameters:
      - name: examId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Exam Timetable]
      summary: Full schedule of an exam
      description: Ordered by date then time.
      operationId: getExamTimetablesByExamId
      responses:
        '200':
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExamTimetableResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Marks — workflow
  # ==================================================================
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      required: false
      description: '`property,(asc|desc)`; repeatable.'
      schema:
        type: array
        items: { type: string }
      example: ['createdAt,desc']
...
    AcademicYearIdFilter:
      name: academicYearId
      in: query
      required: false
      schema: { type: string, format: uuid }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the login endpoint. Authentication is still being built, so
        endpoints are currently open and role rules are recorded as TODOs beside each
        controller method.
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: A referenced record does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 404
            error: Not Found
            message: 'Exam not found with id: 9f1c8b2e-...'
            path: /api/v1/exams/9f1c8b2e-...
    ValidationError:
      description: Bean validation failure; offending fields are listed in `fieldErrors`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 400
            error: Bad Request
            message: Validation failed
            path: /api/v1/exams
            fieldErrors:
              examName: Exam name is required
              actorId: Acting admin id is required
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ExamStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    SubjectType:
      type: string
      enum: [COMPULSORY, ELECTIVE]

    # ---------------- shared summaries ----------------
    ActorSummary:
      type: object
      description: A person who created or last changed a record.
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: R. Iyer }
        role:
          type: string
          description: Designation, e.g. PRINCIPAL or SENIOR_TEACHER.
          example: SENIOR_TEACHER
    ClassSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Grade 10 }
    SubjectSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Mathematics }
        code: { type: string, example: MATH001 }
        type: { $ref: '#/components/schemas/SubjectType' }
    AcademicYearSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string, example: 2025-2026 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
...
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Present only for bean validation failures.
      required: [timestamp, status, error, message, path]
    PageMetadata:
      type: object
      properties:
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        number: { type: integer, description: Zero-based page index, example: 0 }
        size: { type: integer, example: 20 }
        numberOfElements: { type: integer, example: 20 }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }

    # ---------------- exam ----------------
    ExamSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Term 1 Examination }
        status: { $ref: '#/components/schemas/ExamStatus' }
    ExamRequest:
      type: object
      required: [academicYearId, examName, actorId]
      properties:
        academicYearId: { type: string, format: uuid }
        examName: { type: string, maxLength: 255, example: Term 1 Examination }
        status:
          allOf: [{ $ref: '#/components/schemas/ExamStatus' }]
          description: Optional; defaults to DRAFT. Ignored on update.
        actorId:
          type: string
          format: uuid
          description: |
            Admin staff performing the action — becomes `createdBy` on create and
            `updatedBy` on update. Removed once authentication lands.
    ExamResponse:
      type: object
      properties:
        examId: { type: string, format: uuid }
        examName: { type: string }
        status: { $ref: '#/components/schemas/ExamStatus' }
        academicYear: { $ref: '#/components/schemas/AcademicYearSummary' }
        createdBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        academicYearId:
          type: string
          format: uuid
          deprecated: true
          description: Use `academicYear`.
        createdById:
          type: string
          format: uuid
          deprecated: true
          description: Use `createdBy`.
...
    ExamSubjectResponse:
      type: object
      properties:
        examSubjectId: { type: string, format: uuid }
        maxMarks: { type: integer, example: 100 }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        examId: { type: string, format: uuid, deprecated: true }
        subjectId: { type: string, format: uuid, deprecated: true }
        classId: { type: string, format: uuid, deprecated: true }
...
    ExamTimetableResponse:
      type: object
      properties:
        examTimetableId: { type: string, format: uuid }
        examDate: { type: string, format: date }
        examTime: { type: string, example: '09:30:00' }
        maxMarks: { type: integer }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        examSubjectId: { type: string, format: uuid, deprecated: true }
...
    ExamSubjectProgress:
      type: object
      properties:
        examSubjectId: { type: string, format: uuid }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        maxMarks: { type: integer }
        scheduled: { type: boolean }
        examDate: { type: string, format: date, nullable: true }
        examTime: { type: string, nullable: true, example: '09:30:00' }
        marksEntered: { type: integer, format: int64, example: 37 }
    ExamOverviewResponse:
      type: object
      properties:
        examId: { type: string, format: uuid }
        examName: { type: string }
        status: { $ref: '#/components/schemas/ExamStatus' }
        academicYear: { $ref: '#/components/schemas/AcademicYearSummary' }
        createdBy: { $ref: '#/components/schemas/ActorSummary' }
        totalSubjects: { type: integer, example: 6 }
        scheduledSubjects: { type: integer, example: 4 }
        readyToPublish: { type: boolean }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/ExamSubjectProgress' }
    PublishPreviewResponse:
      type: object
      properties:
        examId: { type: string, format: uuid }
        examName: { type: string }
        readyToPublish: { type: boolean }
        totalSubjects: { type: integer }
        scheduledSubjects: { type: integer }
        blockers:
          type: array
          description: Conditions that will cause publish to be rejected.
          items: { type: string }
          example: ['2 subject(s) have not been scheduled in the timetable.']
        warnings:
          type: array
          description: Worth surfacing, but not blocking.
          items: { type: string }
          example: ['3 subject(s) have no marks entered yet.']
...
    PageExamResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ExamResponse' }
...